requirements.txt
asgiref==3.5.2
Django==4.0.6
django-ckeditor==6.4.2
django-ckeditor-5==0.1.6
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-taggit==3.0.0
django-taggit-serializer==0.1.7
django-tastypie==0.14.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==13.15.1
Pillow==9.2.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
